+++ loan-service/src/main/java/com/example/loanservice/repository/LoanRepository.java
@@ -3,5 +3,5 @@ package com.example.loanservice.repository;
 import com.example.loanservice.entity.Loan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoanRepository extends JpaRepository<Loan, Integer> {
+public interface LoanRepository extends JpaRepository<Loan, Long> {
 }
